Fixed header logo responsiveness
